+++ static/components/SessionInfo.js
@@ -5,7 +5,6 @@ var SubStepsActions = require('../actions/SubStepsActions').SubStepsActions;
 var SessionsStore = require('../stores/SessionsStore').SessionsStore;
 var StepsStore = require('../stores/StepsStore').StepsStore;
 var SubStepsStore = require('../stores/SubStepsStore').SubStepsStore;
-var formatDateTime = require('../utils/Utils').formatDateTime;
 var calculatePadding = require('../utils/Utils').calculatePadding;
 var statusIcon = require('./Session').statusIcon;
 
@@ -17,60 +16,24 @@ var SessionInfo = React.createClass({
 
     getInitialState: function() {
         return {
-            session: '',
-            first_step: '',
-            first_sub_step: ''
+            session: ''
         };
     },
 
     _onChangeSession: function() {
-        var _state = SessionsStore.getState(),
-            first_step = this.state.first_step;
+        var _state = SessionsStore.getState();
 
         this.setState({
-            session: _state.session,
-            first_step: first_step
-        });
-    },
-
-    _onChangeSteps: function() {
-        var _state = StepsStore.getState(),
-            session = this.state.session;
-
-        if (_state.steps.length >= 1) {
-            this.setState({
-                session: session,
-                first_step: _state.steps[0]
-            });
-
-            if (!_state.first_sub_step && this.state.first_step) {
-                SubStepsActions.get_sub_steps_for_step(this.state.first_step.id);
-            }
-        }
-    },
-
-    _onChangeSubSteps: function() {
-        var _state = SubStepsStore.getState(),
-            session = this.state.session,
-            first_step = this.state.first_step;
-
-        this.setState({
-            session: session,
-            first_step: first_step,
-            first_sub_step: _state.first_sub_step[0]
+            session: _state.session
         });
     },
 
     componentWillUnmount: function() {
         SessionsStore.removeChangeListener(this._onChangeSession);
-        StepsStore.removeChangeListener(this._onChangeSteps);
-        SubStepsStore.removeChangeListener(this._onChangeSubSteps);
     },
 
     componentDidMount: function() {
         SessionsStore.addChangeListener(this._onChangeSession);
-        StepsStore.addChangeListener(this._onChangeSteps);
-        SubStepsStore.addChangeListener(this._onChangeSubSteps);
     },
 
     componentDidUpdate: function () {
@@ -79,95 +42,12 @@ var SessionInfo = React.createClass({
 
     render: function () {
         var session = this.state.session,
-            first_step = this.state.first_step,
-            first_sub_step = this.state.first_sub_step;
-
-        return (
-            <div className="session_info_panel">
-                <InfoPanel session={ session } first_step={ first_step } first_sub_step={ first_sub_step }/>
-                <SessionTabs session={ session }/>
-                <Snippets/>
-            </div>
-        );
-    }
-});
-
-
-var InfoPanel = React.createClass({
-    description: function () {
-        var desc = {
-                "status_message": '',
-                "session_error": '',
-                "started": '',
-                "ended": '',
-                "duration": '',
-                "username": '',
-                "platform": '',
-                "java": '',
-                "browser": '',
-                "selenium": ''
-            },
-            session = this.props.session,
-            first_step = this.props.first_step,
-            first_sub_step = this.props.first_sub_step;
+            session_status_class = "label label-" + statusIcon(session.status);
 
         if (session) {
-            if (session.endpoint_name) {
-                if (session.status == "waiting") {
-                    desc.status_message = "Starting on " + session.endpoint_name + " ...";
-                } else {
-                    desc.status_message = "Started on " + session.endpoint_name;
-                }
-            } else {
-                desc.status_message = "Waiting for endpoint...";
-            }
-
-            if (session.deleted) {
-                desc.ended = formatDateTime(session.deleted);
-            }
-
-            desc.duration = session.duration ? session.duration : '';
-            desc.username = session.username ? session.username : '';
-            desc.started = formatDateTime(session.created);
-            desc.session_error = session.error;
-        }
-
-        if (first_step && first_step.response) {
-            var dc = JSON.parse(first_step.response.body).value;
-            desc.browser = dc.version && dc.browserName ? dc.browserName + " " + dc.version : '';
-
-            if (first_sub_step) {
-                var sub_step_dc = JSON.parse(first_sub_step.response.body).value;
-                desc.java = sub_step_dc.java.version ? 'Java ' + sub_step_dc.java.version : '';
-                desc.platform = dc.platform && sub_step_dc.os.version ? dc.platform + " " + sub_step_dc.os.version : '';
-                desc.selenium = sub_step_dc.build.version ? 'Selenium ' + sub_step_dc.build.version : '';
-            }
-        }
-
-        return desc;
-    },
-
-    errorMsg: function (error) {
-        if (error) {
             return (
-                <div className="row _info_error">
-                    <div className="col-lg-12">
-                        <pre className="alert alert-danger">{ error }</pre>
-                    </div>
-                </div>
-            );
-        }
-    },
-
-    render: function () {
-        var session = this.props.session,
-            session_status_class = "label label-" + statusIcon(session.status),
-            description = this.description();
-
-        return (
-            <div className="session_info panel">
-                <div className="panel-body">
-                    <div className="container-fluid">
+                <div className="session_info_panel">
+                    <div className="session_info panel">
                         <div className="row _info_title">
                             <ul className="list-inline">
                                 <li className="session_column session_status">
@@ -178,28 +58,14 @@ var InfoPanel = React.createClass({
                                 </li>
                             </ul>
                         </div>
-
-                        <div className="row _info_description">
-                            <div className="col-xs-5 grey-block">
-                                <p><strong>{ description.status_message }</strong></p>
-                                <p><strong>{ description.platform }</strong></p>
-                                <p><strong>{ description.browser }</strong></p>
-                                <p><strong>{ description.java }</strong></p>
-                                <p><strong>{ description.selenium }</strong></p>
-                            </div>
-                            <div className="col-xs-5 grey-block">
-                                <p><strong>Owner</strong> { description.username }</p>
-                                <p><strong>Started</strong> { description.started }</p>
-                                <p><strong>Ended</strong> { description.ended }</p>
-                                <p><strong>Duration</strong> { description.duration } sec.</p>
-                            </div>
-                        </div>
-
-                        { this.errorMsg(description.session_error) }
                     </div>
+                    <SessionTabs session={ session }/>
                 </div>
-            </div>
-        )
+            );
+        }
+
+        return null;
+
     }
 });
 
@@ -209,10 +75,14 @@ var SessionTabs = React.createClass({
         if (!session.closed || session.take_screencast) {
             return (
                 <li>
-                    <a data-toggle="tab" href="#screencast" className="screencast">Video</a>
+                    <a data-toggle="tab" href="#screencast" className="screencast" onClick={this.handleClick}>Video</a>
                 </li>
             );
         }
+    },
+
+    handleClick: function () {
+        window.location.hash = $(".nav-tabs .active a")[0].href.split("/").pop();
     },
 
     render: function () {
@@ -220,7 +90,8 @@ var SessionTabs = React.createClass({
         return (
             <div className="session_tabs">
                 <ul id="mtabs" className="nav nav-tabs">
-                    <li className="active"><a data-toggle="tab" href="#steps" className="steps">Steps</a></li>
+                    <li><a data-toggle="tab" href="#description" className="description" onClick={this.handleClick}>Description</a></li>
+                    <li className="active"><a data-toggle="tab" href="#steps" className="steps" onClick={this.handleClick}>Steps</a></li>
                     { this.screencastTab(session) }
                 </ul>
             </div>
@@ -229,19 +100,30 @@ var SessionTabs = React.createClass({
 });
 
 
-var Snippets = React.createClass({
-    expandButton: function () {
-        return (
-            <div className="expand_all_button">
-                <a type="button" className="btn btn-default" href="javascript: expand_all()">Expand all tests</a>
-            </div>
-        );
+var ExpandButton = React.createClass({
+    componentDidMount: function () {
+        if (window.location.hash) {
+            var tab_for_open = window.location.hash.replace("#", "");
+            $("." + tab_for_open).click();
+        }
     },
 
     render: function () {
         return (
+            <div className="expand_all_button">
+                <a type="button" className="btn btn-default"
+                   href="javascript: expand_all()">Expand all tests</a>
+            </div>
+        );
+    }
+});
+
+
+var Snippets = React.createClass({
+    render: function () {
+        return (
             <div className="step_snippet">
-                { document.getElementsByClassName('label_group').length > 0 ? this.expandButton() : '' }
+                { this.props.children }
             </div>
         );
     }
@@ -249,3 +131,5 @@ var Snippets = React.createClass({
 
 
 module.exports.SessionInfo = SessionInfo;
+module.exports.Snippets = Snippets;
+module.exports.ExpandButton = ExpandButton;
