+++ static/js/session.js
@@ -1,59 +1,3 @@
-$(document).ready(function() {
-    $(function() {
-        if (window.location.hash) {
-            var tab_for_open = window.location.hash.replace("#", "");
-            $("." + tab_for_open).click();
-            if (tab_for_open !== "steps") {
-                $(".expand_all_button").css({
-                    'display': 'none'
-                });
-            } else {
-                $(".expand_all_button").css({
-                    'display': 'block'
-                 });
-            }
-        }
-    });
-
-    function change_hash () {
-        window.location.hash = this.href.split("/").pop();
-        var tab_for_open = window.location.hash.replace("#", "");
-        if (tab_for_open !== "steps") {
-             $(".expand_all_button").css({
-                'display': 'none'
-             });
-        } else {
-            $(".expand_all_button").css({
-                'display': 'block'
-             });
-        }
-    }
-
-    $(function() {
-        var screencast = $("#screencast").children()[1],
-            info_height = $('.session_info_panel').height(),
-            window_height = $(window).height() - 100,
-            new_height = window_height - info_height;
-
-        if (screencast) {
-            if (new_height > 600) {
-                screencast.height = '600';
-            } else if (new_height < 400) {
-                screencast.height = '400';
-            } else {
-                screencast.height = new_height;
-            }
-        }
-    });
-
-    $(function() {
-        var tabs = $(".nav-tabs").children().children();
-        tabs.each(function(i, item) {
-            item.onclick = change_hash
-        });
-    });
-});
-
 function expand_all() {
     var label_groups = $('.label_group'),
         button = $('.btn');
