+++ static/js/vnc.js
@@ -141,46 +141,3 @@ function connect(host, port) {
     rfb.connect(host, port, password, path);
     $(".play").addClass("_playing");
 }
-
-$(document).ready(function() {
-    $(function() {
-        var play = $(".play"),
-            play_icon = $('.glyphicon-play'),
-            tabs_height = $('.session_tabs').height(),
-            info_title_height = $('.info_title').height(),
-            window_height = $(window).height() - 100,
-            new_height = window_height - tabs_height - info_title_height;
-
-        if (new_height > 600) {
-            screencast_size = [800, 600];
-            play.css({
-                'width': 800,
-                'height': 600
-            });
-            play_icon.css({
-                'top': 200,
-                'font-size': 200
-            });
-        } else if (new_height < 400) {
-            screencast_size = [600, 400];
-            play.css({
-                'width': 600,
-                'height': 400
-            });
-            play_icon.css({
-                'top': 150,
-                'font-size': 150
-            });
-        } else {
-            screencast_size = [new_height * 4 / 3, new_height];
-            play.css({
-                'width': new_height * 4 / 3,
-                'height': new_height
-            });
-            play_icon.css({
-                'top': new_height * 4 / 10,
-                'font-size': new_height / 4
-            });
-        }
-    });
-});
